refactor: add custom hook for video and score timer
